+++ packages/gov-ui-kit/CHANGELOG.md
@@ -13,7 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
-- Update style of `ProposalActionItem` module component, render link to block explorer for target address
+- Update `<DialogHeader />` core component to not truncate description strings
+- Update style of `<ProposalActionItem />` module component, render link to block explorer for target address
 - Rename `tokenPrice` property to `amountUsd` on `<TransactionDataListItem />` module component
 - Update workflow to deploy Storybook application to Github pages
 - Update minor and patch NPM dependencies