+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.62] - 2025-01-08
+
 ### Fixed
 
 - Fix `AssetTransfer` module component to not link to the block explorer for native transfers